btf_encoder: use bitfield to control var encoding
We will need more granularity in the future, in order to add support for
encoding global variables as well. So replace the skip_encoding_vars
boolean with a flag variable named "encode_vars". There is currently
only one bit specified, and it is set when percpu variables should be
emitted.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>